refactor(settings): move settings page to views, checkout to billing
