+++ src/views/settings/account-access/success-panel.vue
