+++ src/views/settings/danger-delete-account-button.vue
