+++ src/views/settings/tab-app/accessibility-section.vue
