+++ src/views/settings/tab-danger-zone/index.vue
