+++ src/views/settings/tab-review-preferences/ratings-section.vue
