+++ src/views/settings/tab-review-preferences/review-ratings-toggle.vue
