+++ src/views/settings/tab-subscription/plan-pill.vue
