+++ src/views/settings/use-account-access-click.ts
